[DotNetCore] Rely on MSBuild to determine transitve project references
Removed code from the SdkProjectExtension which determined transitive
project references. This information is available from MSBuild if
a NuGet restore has been run.

Modified the transitive project reference tests so that they
run a NuGet restore

The ThreeProjectReferencesAndReferenceOutputAssemblyIsFalse test was
failing since MSBuild does not include project references that
have ReferenceOutputAssembly set to false. The test was expecting
such a reference to be returned.