[GetToCode] Adds ShowAsync method in NewProjectController with TaskCompletionSource This PR fixes a problem waiting the response from NewProject window probably introduced by 1578ca3#diff-e3e1db8132f317c043c737963f36d108
The problem is now we close the dialog very quickly without wait the creation of the project, then the variable is modified after return the value.

This is necessary to fix, because GTC in case of project creation needs to disable the window to don't allow the user to click buttons meanwhile we open the project.